Move RawImage type to twix
