+++ tools/twix/src/panels/image/mod.rs
@@ -10,12 +10,12 @@ use egui_extras::RetainedImage;
 use log::error;
 use nalgebra::{vector, Similarity2};
 use serde_json::{json, Value};
-use types::image::ImageRaw;
 
 use crate::{
     image_buffer::ImageBuffer,
     nao::Nao,
     panel::Panel,
+    raw_image::RawImage,
     twix_painter::{CoordinateSystem, TwixPainter},
 };
 
@@ -105,7 +105,7 @@ impl ImagePanel {
             .image_buffer
             .get_latest()
             .map_err(|error| eyre!("{error}"))?;
-        let image_raw = bincode::deserialize::<ImageRaw>(&image_data)?;
+        let image_raw = bincode::deserialize::<RawImage>(&image_data)?;
         let image = RetainedImage::from_image_bytes("image", &image_raw.buffer)
             .map_err(|error| eyre!("{error}"))?;
         let image_size = image.size_vec2();
